# src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-billing-preprocessor
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "4.4"
servers:
  - url: http://localhost:53178
    description: Generated server url
...
paths:
...
  /{municipalityId}/jobs/filetransferrer:
    post:
      tags:
        - Jobs
      summary: Triggers job for transferring files with status CREATED or SEND_FAILED
        to final destination
      operationId: transferFiles
      parameters:
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
      responses:
        "202":
          description: Successful Operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/jobs/filecreator:
    post:
      tags:
        - Jobs
      summary: Triggers job for creating files from billing records with status APPROVED
      operationId: createFileEntities
      parameters:
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
      responses:
        "202":
          description: Successful Operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
...
  securitySchemes: {}